HealthBar .2
Added:
New 3d blocks for health bar
Podium
Floor(For gun retrival)

Changed: organized the conents into empty game objects so waves and waypoints are in their own sections

Currently Trying to add a working health bar using a 3d block. Right now the current slider updates the locaiton of the block instead of the scale of the block, However I am on the right path to add the block.